# docs/openapi.yml
openapi: 3.0.1
info:
  title: Opdex Platform API
  version: 1.0.0
  contact:
    name: Opdex Contributors
    url: https://github.com/Opdex/opdex-v1-api
  description: The Opdex Platform Web API is an interface for querying information from the Opdex contract indexer and quoting Opdex smart contract transactions. The Opdex contract indexer is a job that continually runs and stores relevant broadcast transaction data, in a way that represents the current known state of the Opdex protocol at the present time.
  license:
    name: MIT
    url: https://github.com/Opdex/opdex-v1-api/blob/main/LICENSE
servers:
  - url: https://test-api.opdex.com
...
components:
  securitySchemes:
    opdexAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authorize with a bearer token obtained from SignalR
  responses:
...
    MarketOrTokenNotFound:
      description: Market or token not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/problemDetails"
          example:
            type: https://httpstatuses.com/404
            title: Not Found
            detail: Token not found.
            status: 404
            extensions:
              traceId: 00-00000000000000000000000000000000-0000000000000000-00
...
    TooManyRequests:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 30
          description: Indicates how many seconds to wait before making a follow-up request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/problemDetails"
          example:
            type: https://httpstatuses.com/429
            title: Too Many Requests
            detail: "Quota exceeded. Maximum allowed: 500 per 60s. Please try again in 30 second(s)."
            status: 429
            extensions:
              traceId: 00-00000000000000000000000000000000-0000000000000000-00
...
  schemas:
...
    validationProblemDetails:
      allOf:
        - $ref: "#/components/schemas/problemDetails"
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
                  description: Validation error message
                minItems: 1
              description: Contains invalid properties and associated error messages
    address:
      type: string
      description: An address on the Cirrus network
      minLength: 30
      maxLength: 42
      pattern: ^[a-km-zA-HJ-NP-Z1-9]$
...
    marketTokenResponse:
      type: object
      description: Details of a token within a market
      properties:
        allOf:
          $ref: "#/components/schemas/tokenResponse"
          type: object
          properties:
            market:
              $ref: "#/components/schemas/address"
              description: Address of the market
            liquidityPool:
              $ref: "#/components/schemas/address"
              description: Address of the liquidity pool
...
paths:
...
  /markets/{market}/tokens/{token}:
    get:
      tags:
        - Market Tokens
      summary: Get Market Token
      description: Retrieves details of a token, that is tracked within a specified market.
      operationId: getMarketToken
      parameters:
        - name: market
          in: path
          description: Address of the market
          required: true
          schema:
            $ref: "#/components/schemas/address"
          example: t7RorA7xQCMVYKPM1ibPE1NSswaLbpqLQb
        - name: token
          in: path
          description: Address of the token
          required: true
          schema:
            $ref: "#/components/schemas/address"
          example: tGSk2dVENuqAQ2rNXbui37XHuurFCTqadD
      responses:
        200:
          description: Market token found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/marketTokenResponse"
              example:
                market: t7RorA7xQCMVYKPM1ibPE1NSswaLbpqLQb
                liqudityPool: tMdZ2UfwJorAyErDvqNdVU8kmiLaykuE5L
                address: tGSk2dVENuqAQ2rNXbui37XHuurFCTqadD
                name: Bitcoin (Wrapped)
                symbol: TBTC
                decimals: 8
                sats: 100000000
                totalSupply: "21000000.00000000"
                attributes:
                  - NonProvisional
                summary:
                  priceUsd: 54827.48328399
                  dailyPriceChangePercent: 2.42580923
                  modifiedBlock: 2963822
        400:
          description: The request is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/validationProblemDetails"
              example:
                type: https://httpstatuses.com/400
                title: One or more validation errors occurred.
                status: 400
                errors:
                  token:
                    - Invalid address.
                extensions:
                  traceId: 00-00000000000000000000000000000000-0000000000000000-00
        404:
          $ref: "#/components/responses/MarketOrTokenNotFound"
        429:
          $ref: "#/components/responses/TooManyRequests"
        500:
          $ref: "#/components/responses/InternalServerError"
  /markets/{market}/tokens/{token}/history:
    get:
      tags:
        - Market Tokens
      summary: Get Market Token History
      description: Retrieves historical pricing data for a token tracked within a specified market. This is a [paginated endpoint](https://docs.opdex.com/reference/using-the-opdex-platform-api#paginated-endpoints), so a request will retrieve a single page of results. It is recommended to supply filters to your query, to reduce the total number of requests that you may have to make.
      operationId: getMarketTokenHistory
      parameters:
        - name: market
          in: path
          description: Address of the market
          required: true
          schema:
            $ref: "#/components/schemas/address"
          example: t7RorA7xQCMVYKPM1ibPE1NSswaLbpqLQb
        - name: token
          in: path
          description: Address of the token
          required: true
          schema:
            $ref: "#/components/schemas/address"
          example: tGSk2dVENuqAQ2rNXbui37XHuurFCTqadD
        - name: interval
          in: query
          description: Time range between each snapshot
          schema:
            $ref: "#/components/schemas/interval"
          example: 1D
        - name: startDateTime
          in: query
          description: Start time for which to retrieve snapshots
          required: true
          schema:
            type: string
            format: date-time
          example: 2022-01-01T00:00:00Z
        - name: endDateTime
          in: query
          description: End time for which to retrieve snapshots
          required: true
          schema:
            type: string
            format: date-time
          example: 2022-12-31T23:59:59
        - name: direction
          in: query
          description: Order direction of the results
          schema:
            $ref: "#/components/schemas/sortDirection"
          example: DESC
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 750
          example: 100
        - name: cursor
          in: query
          description: Reference of the requested page, returned by a previous call
          schema:
            type: string
      responses:
        200:
          description: Market token snapshot results found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tokenSnapshotsResponse"
              example:
                properties:
                  results:
                    - price:
                        open: 54274.35344481
                        high: 55172.00419378
                        low: 54239.99642473
                        close: 55000.14239378
                      timestamp: 2022-01-04T00:00:00Z
                    - price:
                        open: 52159.53226778
                        high: 54463.33523111
                        low: 52159.53226778
                        close: 54274.35344481
                      timestamp: 2022-01-03T00:00:00Z
                    - price:
                        open: 53163.84322264
                        high: 54339.99180028
                        low: 52101.00022178
                        close: 52159.53226778
                      timestamp: 2022-01-02T00:00:00Z
                    - price:
                        open: 54293.91289928
                        high: 54333.91319534
                        low: 52983.129000420
                        close: 53163.84322264
                      timestamp: 2022-01-01T00:00:00Z
                  paging: {}
        400:
          description: The request is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/validationProblemDetails"
              example:
                type: https://httpstatuses.com/400
                title: One or more validation errors occurred.
                status: 400
                errors:
                  interval:
                    - Filters cannot be provided alongside cursor.
                extensions:
                  traceId: 00-00000000000000000000000000000000-0000000000000000-00
        404:
          $ref: "#/components/responses/MarketOrTokenNotFound"
        429:
          $ref: "#/components/responses/TooManyRequests"
        500:
          $ref: "#/components/responses/InternalServerError"
...
  /tokens:
    get:
      tags:
        - Tokens
      summary: Get Tokens
      description: Retrieves details of all tracked tokens across markets. This is a [paginated endpoint](https://docs.opdex.com/reference/using-the-opdex-platform-api#paginated-endpoints), so a request will retrieve a single page of results. It is recommended to supply filters to your query, to reduce the total number of requests that you may have to make.
      operationId: getTokens
      parameters:
        - name: tokenType
          in: query
          description: Token type filter
          schema:
            $ref: "#/components/schemas/tokenType"
          example: Provisional
        - name: tokens
          in: query
          description: Specific tokens to include
          schema:
            type: array
            items:
              $ref: "#/components/schemas/address"
              description: Address of a token
        - name: keyword
          in: query
          description: Keyword search against token address, name and ticker symbol
          schema:
            type: string
        - name: orderBy
          in: query
          description: Property by which to sort results
          schema:
            $ref: "#/components/schemas/tokenOrderByType"
        - name: direction
          in: query
          description: Order direction of the results
          schema:
            $ref: "#/components/schemas/sortDirection"
          example: DESC
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 50
          example: 10
        - name: cursor
          in: query
          description: Reference of the requested page, returned by a previous call
          schema:
            type: string
      responses:
        200:
          description: Token results found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tokensResponse"
              example:
                results:
                  - address: tGSk2dVENuqAQ2rNXbui37XHuurFCTqadD
                    name: Bitcoin (Wrapped)
                    symbol: TBTC
                    decimals: 8
                    sats: 100000000
                    totalSupply: "21000000.00000000"
                    attributes:
                      - NonProvisional
                    summary:
                      priceUsd: 54911.53361129
                      dailyPriceChangePercent: 2.37290288
                      modifiedBlock: 2963825
                  - address: tF83sdXXt2nTkL7UyEYDVFMK4jTuYMbmR3
                    name: ETH (Wrapped)
                    symbol: TETH
                    decimals: 18
                    sats: 1000000000000000000
                    totalSupply: "100000000.000000000000000000"
                    attributes:
                      - NonProvisional
                    summary:
                      priceUsd: 3827.34621953
                      dailyPriceChangePercent: 2.01006408
                      modifiedBlock: 2963712
                paging:
                  next: ZGlyZWN0aW9uOkRFU0M7bGltaXQ6MTA7cGFnaW5nOkZvcndhcmQ7cHJvdmlzaW9uYWw6QWxsO2tleXdvcmQ6O29yZGVyQnk6UHJpY2VVc2Q7cG9pbnRlcjpLREF1TkRBMk1ETXpPREFzSURRcDs=
        400:
          description: The request is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/validationProblemDetails"
              example:
                type: https://httpstatuses.com/400
                title: One or more validation errors occurred.
                status: 400
                errors:
                  tokenType:
                    - The value 'New' is not valid for TokenType.
                    - Filters cannot be provided alongside cursor.
                extensions:
                  traceId: 00-00000000000000000000000000000000-0000000000000000-00
        429:
          $ref: "#/components/responses/TooManyRequests"
        500:
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Tokens
      summary: Add Token
      description: Adds an SRC token to the Opdex indexer, so that it can be tracked and used within markets.
      operationId: addToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/addTokenRequest"
            example:
              tokenAddress: tF83sdXXt2nTkL7UyEYDVFMK4jTuYMbmR3
      security:
        - opdexAuth: []
      responses:
        201:
          description: Token was added to the indexer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tokenResponse"
              example:
                address: tA88mLAYt3gTp5eqFR6srceAPrTNf4tdFy
                name: LTC (Wrapped)
                symbol: xLTC
                decimals: 8
                sats: 100000000
                totalSupply: "84000000.00000000"
                attributes: []
          headers:
            Location:
              schema:
                type: string
                example: /tokens/tF83sdXXt2nTkL7UyEYDVFMK4jTuYMbmR3
              description: Request location to retrieve details of the added token
        303:
          description: Token is already indexed
          headers:
            Location:
              schema:
                type: string
                example: /tokens/tF83sdXXt2nTkL7UyEYDVFMK4jTuYMbmR3
              description: Request location to retrieve details of the added token
        400:
          description: The request is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/validationProblemDetails"
              example:
                type: https://httpstatuses.com/400
                title: One or more validation errors occurred.
                status: 400
                errors:
                  tokenAddress:
                    - Invalid address.
                extensions:
                  traceId: 00-00000000000000000000000000000000-0000000000000000-00
        401:
          $ref: "#/components/responses/Unauthorized"
        429:
          $ref: "#/components/responses/TooManyRequests"
        500:
          $ref: "#/components/responses/InternalServerError"
  /tokens/{token}:
    get:
      tags:
        - Tokens
      summary: Get Token
      description: Retrieves details of a token tracked by the Opdex indexer.
      operationId: getToken
      parameters:
        - name: token
          in: path
          description: Address of the token
          required: true
          schema:
            $ref: "#/components/schemas/address"
          example: tGSk2dVENuqAQ2rNXbui37XHuurFCTqadD
      responses:
        200:
          description: Token details found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tokenResponse"
              example:
                address: tGSk2dVENuqAQ2rNXbui37XHuurFCTqadD
                name: Bitcoin (Wrapped)
                symbol: TBTC
                decimals: 8
                sats: 100000000
                totalSupply: "21000000.00000000"
                attributes:
                  - NonProvisional
                summary:
                  priceUsd: 54911.53361129
                  dailyPriceChangePercent: 2.37290288
                  modifiedBlock: 2963825
        400:
          description: The request is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/validationProblemDetails"
              example:
                type: https://httpstatuses.com/400
                title: One or more validation errors occurred.
                status: 400
                errors:
                  token:
                    - Invalid address.
                extensions:
                  traceId: 00-00000000000000000000000000000000-0000000000000000-00
        404:
          $ref: "#/components/responses/TokenNotFound"
        429:
          $ref: "#/components/responses/TooManyRequests"
        500:
          $ref: "#/components/responses/InternalServerError"
  /tokens/{token}/history:
    get:
      tags:
        - Tokens
      summary: Get Token History
      description: Retrieves historical pricing data for a token tracked by the Opdex indexer. This is a [paginated endpoint](https://docs.opdex.com/reference/using-the-opdex-platform-api#paginated-endpoints), so a request will retrieve a single page of results. It is recommended to supply filters to your query, to reduce the total number of requests that you may have to make.
      operationId: getTokenHistory
      parameters:
        - name: token
          in: path
          description: Address of the token
          required: true
          schema:
            $ref: "#/components/schemas/address"
          example: tGSk2dVENuqAQ2rNXbui37XHuurFCTqadD
        - name: interval
          in: query
          description: Time range between each snapshot
          schema:
            $ref: "#/components/schemas/interval"
          example: 1D
        - name: startDateTime
          in: query
          description: Start time for which to retrieve snapshots
          required: true
          schema:
            type: string
            format: date-time
          example: 2022-01-01T00:00:00Z
        - name: endDateTime
          in: query
          description: End time for which to retrieve snapshots
          required: true
          schema:
            type: string
            format: date-time
          example: 2022-12-31T23:59:59
        - name: direction
          in: query
          description: Order direction of the results
          schema:
            $ref: "#/components/schemas/sortDirection"
          example: DESC
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 750
          example: 100
        - name: cursor
          in: query
          description: Reference of the requested page, returned by a previous call
          schema:
            type: string
      responses:
        200:
          description: Token snapshot results found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tokenSnapshotsResponse"
              example:
                properties:
                  results:
                    - price:
                        open: 54193.34115385
                        high: 55199.32820982
                        low: 54187.23589023
                        close: 55063.99234829
                      timestamp: 2022-01-04T00:00:00Z
                    - price:
                        open: 52402.45948025
                        high: 54362.13294822
                        low: 52351.34292224
                        close: 54193.34115385
                      timestamp: 2022-01-03T00:00:00Z
                    - price:
                        open: 53135.84898883
                        high: 54339.34298819
                        low: 52217.00424382
                        close: 52402.45948025
                      timestamp: 2022-01-02T00:00:00Z
                    - price:
                        open: 54284.23999384
                        high: 54300.22222003
                        low: 53025.42828821
                        close: 53135.84898883
                      timestamp: 2022-01-01T00:00:00Z
                  paging: {}
        400:
          description: The request is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/validationProblemDetails"
              example:
                type: https://httpstatuses.com/400
                title: One or more validation errors occurred.
                status: 400
                errors:
                  token:
                    - Invalid address.
                extensions:
                  traceId: 00-00000000000000000000000000000000-0000000000000000-00
        404:
          $ref: "#/components/responses/TokenNotFound"
        429:
          $ref: "#/components/responses/TooManyRequests"
        500:
          $ref: "#/components/responses/InternalServerError"
  /tokens/{token}/approve:
    post:
      tags:
        - Tokens
      summary: Build Approve Allowance Transaction Quote
      description: Builds a quote for an allowance approval transaction. The quote can be broadcast by a Stratis Transaction Handoff Broadcastor. See the [specification](https://github.com/Opdex/STHS) for further details.
      operationId: buildApproveAllowanceQuote
      parameters:
        - name: token
          in: path
          description: Address of the token
          required: true
          schema:
            $ref: "#/components/schemas/address"
          example: tGSk2dVENuqAQ2rNXbui37XHuurFCTqadD
      security:
        - opdexAuth: []
      responses:
        200:
          description: Built approve allowance transaction quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/transactionQuoteResponse"
              example:
                result: true
                gasUsed: 12923
                events:
                  - owner: tQ9RukZsB6bBsenHnGSo1q69CJzWGnxohm
                    spender: tAFxpxRdcV9foADqD6gK3c8sY5MeANzFp5
                    amount: "0.020000000000000000"
                    eventType: ApprovalEvent
                    contract: tF83sdXXt2nTkL7UyEYDVFMK4jTuYMbmR3
                    sortOrder: 0
                request: eyJzZW5kZXIiOiJ0UTlSdWtac0I2YkJzZW5IbkdTbzFxNjlDSnpXR254b2htIiwidG8iOiJ0Rjgzc2RYWHQyblRrTDdVeUVZRFZGTUs0alR1WU1ibVIzIiwiYW1vdW50IjoiMCIsIm1ldGhvZCI6IkFwcHJvdmUiLCJwYXJhbWV0ZXJzIjpbeyJsYWJlbCI6IlNwZW5kZXIiLCJ2YWx1ZSI6IjkjdDhYcEgxcE5ZRGdDbnFrOTFaUUtMZ3BVVmVKN1htb21MVCJ9LHsibGFiZWwiOiJDdXJyZW50IEFsbG93YW5jZSIsInZhbHVlIjoiMTIjMjAwMDAwMDAwMDAwMDAwMCJ9LHsibGFiZWwiOiJOZXcgQWxsb3dhbmNlIiwidmFsdWUiOiIxMiMyMDAwMDAwMDAwMDAwMDAwMCJ9XSwiY2FsbGJhY2siOiJodHRwczovL3Rlc3QtYXBpLm9wZGV4LmNvbS90cmFuc2FjdGlvbnMifQ==
        400:
          description: The request is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/validationProblemDetails"
              example:
                type: https://httpstatuses.com/400
                title: One or more validation errors occurred.
                status: 400
                errors:
                  token:
                    - Invalid address.
                extensions:
                  traceId: 00-00000000000000000000000000000000-0000000000000000-00
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/TokenNotFound"
        429:
          $ref: "#/components/responses/TooManyRequests"
        500:
          $ref: "#/components/responses/InternalServerError"
  /tokens/{token}/distribute:
    post:
      tags:
        - Tokens
      summary: Build Distribute Tokens Transaction Quote
      description: Builds a quote for a transaction to distribute governance tokens, to the mining governance and vault. The quote can be broadcast by a Stratis Transaction Handoff Broadcastor. See the [specification](https://github.com/Opdex/STHS) for further details.
      operationId: buildDistributeTokensQuote
      parameters:
        - name: token
          in: path
          description: Address of the governance token
          required: true
          schema:
            $ref: "#/components/schemas/address"
          example: tBeY2UAVsbHoS9jwEnT2kMmRsJwUHK1j1L
      security:
        - opdexAuth: []
      responses:
        200:
          description: Built distribute tokens transaction quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/transactionQuoteResponse"
              example:
                gasUsed: 99064
                events:
                  - vaultAmount: "100000000.00000000"
                    miningGovernanceAmount: "300000000.00000000"
                    periodIndex: 0
                    totalSupply: "400000000.00000000"
                    nextDistributionBlock: 4629775
                    eventType: DistributionEvent
                    contract: "tBeY2UAVsbHoS9jwEnT2kMmRsJwUHK1j1L"
                    sortOrder: 0
                request: eyJzZW5kZXIiOiJ0SFlIZW03Y0xLZ29Ma2ViNzkyeW40V2F5cUt6THJqSmFrIiwidG8iOiJ0QmVZMlVBVnNiSG9TOWp3RW5UMmtNbVJzSndVSEsxajFMIiwiYW1vdW50IjoiMCIsIm1ldGhvZCI6IkRpc3RyaWJ1dGUiLCJjYWxsYmFjayI6Imh0dHBzOi8vdGVzdC1hcGkub3BkZXguY29tL3RyYW5zYWN0aW9ucyJ9
        400:
          description: The request is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/validationProblemDetails"
              example:
                type: https://httpstatuses.com/400
                title: One or more validation errors occurred.
                status: 400
                errors:
                  token:
                    - Invalid address.
                extensions:
                  traceId: 00-00000000000000000000000000000000-0000000000000000-00
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/TokenNotFound"
        429:
          $ref: "#/components/responses/TooManyRequests"
        500:
          $ref: "#/components/responses/InternalServerError"
...